test(azure): split responses transformation tests
